cec: remove the turn-on logic added before [1/1]
PD#SWPL-22291

Problem:
In OTT-4363, the code of turning on tv when the
playback allocates its logical address is added.
This will cause the tv to wake up and even set
active source even though it's not.

Solution:
Tv could be turned on by playback's OneTouchPlayAction,
and controlled by the switch of one touch play, but
should not just do this at any time.

Verify:
ampere

Change-Id: I5fb5a148b68b1e87a7703b32ec369e892ab9ee3d
Signed-off-by: an.xi <an.xi@amlogic.com>